# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loyalty points & marketplace platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  role: UserRole;
  pointsBalance: number;
  createdAt: string;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

export const ProductType = {
  item: "item",
  coupon: "coupon",
} as const;

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  active: "active",
  sold: "sold",
  draft: "draft",
} as const;

export interface Product {
  id: number;
  sellerId: number;
  sellerName: string;
  /** @nullable */
  sellerAvatarUrl?: string | null;
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  type: ProductType;
  pointPrice: number;
  /** @nullable */
  cashPriceCents?: number | null;
  stock: number;
  status: ProductStatus;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  couponDiscountPct?: number | null;
  /** @nullable */
  couponExpiresAt?: string | null;
  createdAt: string;
}

export type NewProductType =
  (typeof NewProductType)[keyof typeof NewProductType];

export const NewProductType = {
  item: "item",
  coupon: "coupon",
} as const;

export interface NewProduct {
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  type: NewProductType;
  pointPrice: number;
  /** @nullable */
  cashPriceCents?: number | null;
  stock: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  couponDiscountPct?: number | null;
  /** @nullable */
  couponExpiresAt?: string | null;
}

export type ProductUpdateStatus =
  (typeof ProductUpdateStatus)[keyof typeof ProductUpdateStatus];

export const ProductUpdateStatus = {
  active: "active",
  sold: "sold",
  draft: "draft",
} as const;

export interface ProductUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  category?: string;
  pointPrice?: number;
  /** @nullable */
  cashPriceCents?: number | null;
  stock?: number;
  status?: ProductUpdateStatus;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  purchase: "purchase",
  transfer: "transfer",
  exchange: "exchange",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  /** @nullable */
  buyerId?: number | null;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  sellerId?: number | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productTitle?: string | null;
  /** @nullable */
  productImageUrl?: string | null;
  pointsAmount: number;
  /** @nullable */
  cashCents?: number | null;
  /** @nullable */
  note?: string | null;
  status: TransactionStatus;
  createdAt: string;
}

export interface PurchaseBody {
  productId: number;
}

export interface TransferBody {
  recipientId: number;
  pointsAmount: number;
  /** @nullable */
  note?: string | null;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  purchase: "purchase",
  transfer: "transfer",
  system: "system",
  marketplace: "marketplace",
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  title: string;
  body: string;
  read: boolean;
  createdAt: string;
}

export interface MarkAllReadResponse {
  updatedCount: number;
}

export interface DashboardSummary {
  pointsBalance: number;
  pointsEarned30d: number;
  pointsSpent30d: number;
  activeListings: number;
  completedPurchases: number;
  couponsOwned: number;
  unreadNotifications: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  purchase: "purchase",
  transfer: "transfer",
  exchange: "exchange",
  listing: "listing",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  subtitle: string;
  /** @nullable */
  pointsAmount?: number | null;
  /** @nullable */
  actorName?: string | null;
  /** @nullable */
  actorAvatarUrl?: string | null;
  createdAt: string;
}

export interface AdminOverview {
  totalUsers: number;
  totalListings: number;
  activeListings: number;
  transactions7d: number;
  transactions30d: number;
  pointsCirculated: number;
  cashVolumeCents: number;
  newUsers7d: number;
}

export interface TopSeller {
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  totalSales: number;
  pointsEarned: number;
}

export type ListUsersParams = {
  search?: string;
};

export type ListProductsParams = {
  type?: ListProductsType;
  category?: string;
  search?: string;
  sellerId?: number;
  status?: ListProductsStatus;
};

export type ListProductsType =
  (typeof ListProductsType)[keyof typeof ListProductsType];

export const ListProductsType = {
  item: "item",
  coupon: "coupon",
} as const;

export type ListProductsStatus =
  (typeof ListProductsStatus)[keyof typeof ListProductsStatus];

export const ListProductsStatus = {
  active: "active",
  sold: "sold",
  draft: "draft",
} as const;

export type ListTransactionsParams = {
  type?: ListTransactionsType;
  limit?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  purchase: "purchase",
  transfer: "transfer",
  exchange: "exchange",
} as const;