# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loyalty points & marketplace platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List marketplace listings
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [item, coupon] }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: sellerId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [active, sold, draft] }
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NewProduct" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single listing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a listing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Remove a listing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # --- Transactions ---
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Product:
      type: object
      properties:
        id: { type: integer }
        sellerId: { type: integer }
        sellerName: { type: string }
        sellerAvatarUrl:
          type: ["string", "null"]
        title: { type: string }
        description: { type: string }
        imageUrl:
          type: ["string", "null"]
        category: { type: string }
        type: { type: string, enum: [item, coupon] }
        pointPrice: { type: integer }
        cashPriceCents:
          type: ["integer", "null"]
        stock: { type: integer }
        status: { type: string, enum: [active, sold, draft] }
        couponCode:
          type: ["string", "null"]
        couponDiscountPct:
          type: ["integer", "null"]
        couponExpiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
      required:
        - id
        - sellerId
        - sellerName
        - title
        - description
        - category
        - type
        - pointPrice
        - stock
        - status
        - createdAt
    NewProduct:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        imageUrl:
          type: ["string", "null"]
        category: { type: string }
        type: { type: string, enum: [item, coupon] }
        pointPrice: { type: integer }
        cashPriceCents:
          type: ["integer", "null"]
        stock: { type: integer }
        couponCode:
          type: ["string", "null"]
        couponDiscountPct:
          type: ["integer", "null"]
        couponExpiresAt:
          type: ["string", "null"]
          format: date-time
      required: [title, description, category, type, pointPrice, stock]
    ProductUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        imageUrl:
          type: ["string", "null"]
        category: { type: string }
        pointPrice: { type: integer }
        cashPriceCents:
          type: ["integer", "null"]
        stock: { type: integer }
        status: { type: string, enum: [active, sold, draft] }
...
    Notification:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string, enum: [purchase, transfer, system, marketplace] }
        title: { type: string }
        body: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, userId, type, title, body, read, createdAt]
...