# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loyalty points & marketplace platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from "./productStatus";
import type { ProductType } from "./productType";

export interface Product {
  id: number;
  sellerId: number;
  sellerName: string;
  /** @nullable */
  sellerAvatarUrl?: string | null;
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  type: ProductType;
  pointPrice: number;
  /** @nullable */
  cashPriceCents?: number | null;
  stock: number;
  status: ProductStatus;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  couponDiscountPct?: number | null;
  /** @nullable */
  couponExpiresAt?: Date | null;
  createdAt: Date;
}